+++ src/session/push.c
@@ -27,7 +27,7 @@ int8_t _z_trigger_push(_z_session_t *zn, _z_n_msg_push_t *push) {
     // TODO check body to know where to dispatch
     _z_slice_t payload = push->_body._is_put ? push->_body._body._put._payload : _z_slice_empty();
     _z_encoding_t encoding = push->_body._is_put ? push->_body._body._put._encoding : _z_encoding_null();
-    int kind = push->_body._is_put ? Z_SAMPLE_KIND_PUT : Z_SAMPLE_KIND_DELETE;
+    size_t kind = push->_body._is_put ? Z_SAMPLE_KIND_PUT : Z_SAMPLE_KIND_DELETE;
 #if Z_FEATURE_SUBSCRIPTION == 1
     ret = _z_trigger_subscriptions(zn, push->_key, payload, encoding, kind, push->_timestamp, push->_qos,
                                    push->_body._body._put._attachment);
